#3935: Move internal/istio up in folder structure to prevent golang-tools build errors.
