+++ cmd/controller/app/BUILD.bazel
@@ -21,9 +21,9 @@ go_library(
         "//pkg/controller/clusterissuers:go_default_library",
         "//pkg/controller/ingress-shim:go_default_library",
         "//pkg/controller/issuers:go_default_library",
+        "//pkg/internal/istio:go_default_library",
         "//pkg/issuer/acme:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
-        "//pkg/issuer/acme/http/internal/istio:go_default_library",
         "//pkg/issuer/ca:go_default_library",
         "//pkg/issuer/selfsigned:go_default_library",
         "//pkg/issuer/vault:go_default_library",
