+++ pkg/controller/acmechallenges/BUILD.bazel
@@ -21,11 +21,11 @@ go_library(
         "//pkg/controller:go_default_library",
         "//pkg/controller/acmechallenges/scheduler:go_default_library",
         "//pkg/feature:go_default_library",
+        "//pkg/internal/istio:go_default_library",
         "//pkg/issuer:go_default_library",
         "//pkg/issuer/acme/dns:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
         "//pkg/issuer/acme/http:go_default_library",
-        "//pkg/issuer/acme/http/internal/istio:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util/feature:go_default_library",
         "@com_github_go_logr_logr//:go_default_library",
